// src/app/public/shared/types/common/learning-form-inner.interface.ts
export interface LearningFormInnerInterface {
  market: string[];
  dateCreation: string;
  stage: string;
  technologies: string[];
  residentList: string[];
  isInnovation: string;
  isMemberMoscow: string;
  isCompanyMSP: string;
  mspCategory: string;
  isPublished: string;
  isSkolkovo: string;
  isStatup: string;
  companyTechnology: string[];
  service: string;
  businessModel: string[];
  fond_type: number;
  result: string;
}