// src/app/public/shared/types/learning-form-request.interface.ts
export interface LearningFormRequestInterface {
  start_up: {
    service: string[];
    found_date: string[];
    market_type: string[];
    tech_type: string[];
    tech_stack: string[];
    techp_residents: string[];
    b_model: string[];
    evo_stage: string[];
    skolk_inno: number[];
    msk_inno: number[];
    nav_stat: number[];
    inno_comp: number[];
    start_up: number[];
    msp: number[];
    msp_cat: string[];
    fond_type: number;
    result: number;
  };
}